Simplifying Geopoint tuple code and renaming NodeOptionsParser to NumberParsers.ts. Moving it to the lib folder
